+++ module-order-service/src/main/java/com/fashionmall/order/controller/BillingKeyController.java
@@ -5,7 +5,7 @@ import com.fashionmall.common.response.PageInfoResponseDto;
 import com.fashionmall.common.util.ApiResponseUtil;
 import com.fashionmall.order.dto.request.BillingKeyRequestDto;
 import com.fashionmall.order.dto.response.UserBillingKeyResponseDto;
-import com.fashionmall.order.service.IamPortService;
+import com.fashionmall.order.service.BillingKeyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,23 +13,23 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class BillingKeyController {
 
-    private final IamPortService iamPortService;
+    private final BillingKeyService billingKeyService;
 
     @PostMapping("/billingKey")
     public CommonResponse<Long> issueBillingKey(@RequestBody BillingKeyRequestDto billingKeyRequestDto) {
 
-        return ApiResponseUtil.success(iamPortService.issueBillingKey(billingKeyRequestDto));
+        return ApiResponseUtil.success(billingKeyService.issueBillingKey(billingKeyRequestDto));
     }
 
     @GetMapping("/billingKey")
     public CommonResponse<PageInfoResponseDto<UserBillingKeyResponseDto>> getBillingKey(@RequestParam(defaultValue = "1") int pageNo,
                                                                                         @RequestParam(defaultValue = "10") int size) {
         Long userId = 1L;
-        return ApiResponseUtil.success(iamPortService.getUserBillingKeyList(userId, pageNo, size));
+        return ApiResponseUtil.success(billingKeyService.getUserBillingKeyList(userId, pageNo, size));
     }
 
     @DeleteMapping("/billingKey/{billingKeyId}")
     public CommonResponse<Long> deleteBillingKey(@PathVariable(value = "billingKeyId") Long billingKeyId) {
-        return ApiResponseUtil.success(iamPortService.deleteBillingKey(billingKeyId));
+        return ApiResponseUtil.success(billingKeyService.deleteBillingKey(billingKeyId));
     }
 }
