+++ module-order-service/src/main/java/com/fashionmall/order/infra/iamPort/dto/TokenRequestDto.java
@@ -1,4 +1,4 @@
-package com.fashionmall.order.dto.request;
+package com.fashionmall.order.infra.iamPort.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
